requirements.txt
chardet==3.0.4
dj-database-url==0.5.0
Django==2.0.5
django-debug-toolbar==1.9.1
django-multiselectfield==0.1.8
djangorestframework==3.8.2
formatizer==0.1.1
future==0.16.0
gunicorn==19.9.0
Pillow==5.1.0
psycopg2==2.6.2
pytz==2018.4
six==1.11.0
sqlparse==0.2.4
whitenoise==4.1
ww==0.2.1
